// libs/services/src/lib/models/enums/PageIndex.ts
export enum PageIndex {
    None = -1,
    Home = 1,
    Team = 2,
    Updates = 3,
    Resources = 4,
    University = 5,

    Catalog = 6,
    CatalogItemDetail = 7,
    CatalogItemViewer = 8,
    PageNotFound = 200
}

export enum ChildPageIndex {
    None = -1,
    Home = 1,
    Catalog = 3,
    About = 2,
    CatalogItemDetail = 4,
    CatalogItemViewer = 5
}